requirements/dev-requirements.txt
-e .
pre-commit
pytest-django
tox

-r ./base-requirements.txt
